add unit icon svg
