+++ src/play/dsstats.play/SpawnPlaybackCanvas.razor.css
@@ -108,6 +108,26 @@
     z-index: 1;
 }
 
+.spawn-playback-unit-cell {
+    display: flex;
+    align-items: center;
+    min-width: 0;
+    gap: 0.35rem;
+}
+
+.spawn-playback-unit-icon {
+    width: 20px;
+    height: 20px;
+    flex: 0 0 20px;
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+}
+
+.spawn-playback-unit-icon svg {
+    display: block;
+}
+
 .spawn-playback-player-name {
     max-width: 13rem;
 }
