+++ src/play/dsstats.play/TypeScript/unitIcons.ts
@@ -1,3 +1,5 @@
+import { TEAM_COLORS } from "./constants";
+import { terranUnits } from "./terranIcons";
 import type {
     CanvasContext,
     UnitIconDefinition,
@@ -5,580 +7,7 @@ import type {
     UnitIconPathCommand,
     UnitIconRenderOptions
 } from "./types";
-
-const terranMarine: UnitIconDefinition = {
-    id: "terran.marine",
-    commander: "terran",
-    aliases: ["Marine", "MarineLightweight"],
-    viewBox: { width: 100, height: 100 },
-    tokens: {
-        badgeFill: "#234A68",
-        badgeGlow: "#5DADEC",
-        badgeStroke: "#B9E1FF",
-
-        armorFill: "#D8E7F0",
-        armorMid: "#9FB3C0",
-        armorShade: "#5E7280",
-        armorDark: "#2E4756",
-
-        visorFill: "#F5D35D",
-        visorShade: "#D9962B",
-
-        redLight: "#FF5A4F",
-        blueLight: "#76D6FF",
-
-        darkStroke: "#102838",
-        rifleFill: "#243946"
-    },
-    layers: [
-        // Badge base
-        // {
-        //     type: "circle",
-        //     cx: 50,
-        //     cy: 50,
-        //     r: 39,
-        //     fill: "badgeFill",
-        //     opacity: 0.98
-        // },
-        // {
-        //     type: "circle",
-        //     cx: 50,
-        //     cy: 50,
-        //     r: 39,
-        //     stroke: "badgeStroke",
-        //     strokeWidth: 4,
-        //     opacity: 0.9
-        // },
-        // {
-        //     type: "circle",
-        //     cx: 50,
-        //     cy: 50,
-        //     r: 33,
-        //     stroke: "badgeGlow",
-        //     strokeWidth: 2,
-        //     opacity: 0.35
-        // },
-
-        // Back shoulder silhouette
-        {
-            type: "path",
-            commands: [
-                ["M", 19, 67],
-                ["C", 21, 55, 29, 47, 38, 47],
-                ["L", 44, 70],
-                ["L", 31, 79],
-                ["C", 24, 78, 20, 74, 19, 67],
-                ["Z"]
-            ],
-            fill: "armorShade",
-            stroke: "darkStroke",
-            strokeWidth: 3,
-            lineJoin: "round"
-        },
-        {
-            type: "path",
-            commands: [
-                ["M", 81, 67],
-                ["C", 79, 55, 71, 47, 62, 47],
-                ["L", 56, 70],
-                ["L", 69, 79],
-                ["C", 76, 78, 80, 74, 81, 67],
-                ["Z"]
-            ],
-            fill: "armorShade",
-            stroke: "darkStroke",
-            strokeWidth: 3,
-            lineJoin: "round"
-        },
-
-        // Helmet dome
-        {
-            type: "path",
-            commands: [
-                ["M", 28, 55],
-                ["C", 28, 37, 37, 25, 50, 23],
-                ["C", 63, 25, 72, 37, 72, 55],
-                ["L", 66, 67],
-                ["L", 34, 67],
-                ["Z"]
-            ],
-            fill: "armorFill",
-            stroke: "darkStroke",
-            strokeWidth: 4,
-            lineJoin: "round"
-        },
-
-        // Helmet side panels
-        {
-            type: "path",
-            commands: [
-                ["M", 29, 50],
-                ["L", 21, 55],
-                ["L", 24, 68],
-                ["L", 34, 66],
-                ["L", 36, 54],
-                ["Z"]
-            ],
-            fill: "armorMid",
-            stroke: "darkStroke",
-            strokeWidth: 3,
-            lineJoin: "round"
-        },
-        {
-            type: "path",
-            commands: [
-                ["M", 71, 50],
-                ["L", 79, 55],
-                ["L", 76, 68],
-                ["L", 66, 66],
-                ["L", 64, 54],
-                ["Z"]
-            ],
-            fill: "armorMid",
-            stroke: "darkStroke",
-            strokeWidth: 3,
-            lineJoin: "round"
-        },
-
-        // Brow plate
-        {
-            type: "path",
-            commands: [
-                ["M", 32, 43],
-                ["C", 38, 36, 44, 33, 50, 33],
-                ["C", 56, 33, 62, 36, 68, 43],
-                ["L", 64, 50],
-                ["C", 59, 46, 55, 44, 50, 44],
-                ["C", 45, 44, 41, 46, 36, 50],
-                ["Z"]
-            ],
-            fill: "armorDark",
-            stroke: "darkStroke",
-            strokeWidth: 2,
-            lineJoin: "round"
-        },
-
-        // Visor
-        {
-            type: "path",
-            commands: [
-                ["M", 34, 50],
-                ["C", 39, 45, 45, 42, 50, 42],
-                ["C", 55, 42, 61, 45, 66, 50],
-                ["L", 62, 59],
-                ["L", 38, 59],
-                ["Z"]
-            ],
-            fill: "visorFill",
-            stroke: "darkStroke",
-            strokeWidth: 3,
-            lineJoin: "round"
-        },
-        {
-            type: "path",
-            commands: [
-                ["M", 39, 55],
-                ["C", 45, 52, 55, 52, 61, 55]
-            ],
-            stroke: "visorShade",
-            strokeWidth: 3,
-            lineCap: "round",
-            opacity: 0.65
-        },
-
-        // Respirator / mouth guard
-        {
-            type: "path",
-            commands: [
-                ["M", 39, 61],
-                ["L", 61, 61],
-                ["L", 58, 72],
-                ["L", 42, 72],
-                ["Z"]
-            ],
-            fill: "armorDark",
-            stroke: "darkStroke",
-            strokeWidth: 3,
-            lineJoin: "round"
-        },
-        {
-            type: "path",
-            commands: [
-                ["M", 44, 64],
-                ["L", 44, 70],
-                ["M", 50, 64],
-                ["L", 50, 71],
-                ["M", 56, 64],
-                ["L", 56, 70]
-            ],
-            stroke: "armorMid",
-            strokeWidth: 2,
-            lineCap: "round"
-        },
-
-        // Chest plate
-        {
-            type: "path",
-            commands: [
-                ["M", 36, 72],
-                ["L", 64, 72],
-                ["L", 70, 84],
-                ["L", 30, 84],
-                ["Z"]
-            ],
-            fill: "armorMid",
-            stroke: "darkStroke",
-            strokeWidth: 3,
-            lineJoin: "round"
-        },
-        {
-            type: "path",
-            commands: [
-                ["M", 43, 75],
-                ["L", 57, 75],
-                ["L", 54, 82],
-                ["L", 46, 82],
-                ["Z"]
-            ],
-            fill: "armorFill",
-            stroke: "darkStroke",
-            strokeWidth: 2,
-            lineJoin: "round"
-        },
-
-        // Helmet lights
-        {
-            type: "circle",
-            cx: 35,
-            cy: 38,
-            r: 3,
-            fill: "redLight",
-            stroke: "darkStroke",
-            strokeWidth: 1.5
-        },
-        {
-            type: "circle",
-            cx: 65,
-            cy: 38,
-            r: 3,
-            fill: "blueLight",
-            stroke: "darkStroke",
-            strokeWidth: 1.5
-        },
-
-        // Small rifle silhouette across lower badge
-        {
-            type: "path",
-            commands: [
-                ["M", 24, 77],
-                ["L", 58, 66],
-                ["L", 61, 70],
-                ["L", 78, 65],
-                ["L", 80, 70],
-                ["L", 63, 75],
-                ["L", 59, 72],
-                ["L", 29, 82],
-                ["Z"]
-            ],
-            fill: "rifleFill",
-            stroke: "darkStroke",
-            strokeWidth: 2,
-            lineJoin: "round",
-            opacity: 0.9
-        },
-
-        // Decorative antenna / comms
-        {
-            type: "path",
-            commands: [
-                ["M", 31, 33],
-                ["L", 22, 22],
-                ["M", 69, 33],
-                ["L", 78, 22],
-                ["M", 45, 25],
-                ["L", 50, 17],
-                ["L", 55, 25]
-            ],
-            stroke: "darkStroke",
-            strokeWidth: 3.5,
-            lineCap: "round",
-            lineJoin: "round"
-        }
-    ]
-};
-
-const zergZergling: UnitIconDefinition = {
-    id: "zerg.zergling",
-    commander: "zerg",
-    aliases: ["Zergling", "ZerglingLightweight"],
-    viewBox: { width: 100, height: 100 },
-    tokens: {
-        skinFill: "#8C5A92",
-        skinMid: "#6F4374",
-        skinShade: "#4F2D56",
-        carapaceFill: "#B78AC8",
-        carapaceShade: "#7C5A8A",
-        clawFill: "#E7D9B8",
-        eyeFill: "#FF5A54",
-        darkStroke: "#2A1630"
-    },
-    layers: [
-        // Tail
-        {
-            type: "path",
-            commands: [
-                ["M", 18, 66],
-                ["C", 10, 64, 8, 58, 14, 54],
-                ["C", 20, 51, 27, 54, 31, 59],
-                ["L", 29, 66],
-                ["Z"]
-            ],
-            fill: "skinShade",
-            stroke: "darkStroke",
-            strokeWidth: 3,
-            lineJoin: "round"
-        },
-
-        // Rear leg
-        {
-            type: "path",
-            commands: [
-                ["M", 30, 66],
-                ["L", 24, 82],
-                ["L", 31, 88],
-                ["L", 41, 76],
-                ["L", 39, 66],
-                ["Z"]
-            ],
-            fill: "skinMid",
-            stroke: "darkStroke",
-            strokeWidth: 3,
-            lineJoin: "round"
-        },
-
-        // Rear claw
-        {
-            type: "path",
-            commands: [
-                ["M", 23, 84],
-                ["L", 18, 91],
-                ["L", 27, 88],
-                ["Z"]
-            ],
-            fill: "clawFill",
-            stroke: "darkStroke",
-            strokeWidth: 2,
-            lineJoin: "round"
-        },
-
-        // Main body
-        {
-            type: "path",
-            commands: [
-                ["M", 27, 61],
-                ["C", 32, 49, 44, 42, 58, 42],
-                ["C", 68, 42, 76, 46, 82, 52],
-                ["C", 84, 58, 81, 64, 74, 67],
-                ["L", 52, 70],
-                ["C", 41, 72, 32, 69, 27, 61],
-                ["Z"]
-            ],
-            fill: "skinFill",
-            stroke: "darkStroke",
-            strokeWidth: 4,
-            lineJoin: "round"
-        },
-
-        // Belly / lower body
-        {
-            type: "path",
-            commands: [
-                ["M", 33, 61],
-                ["C", 40, 66, 49, 68, 60, 67],
-                ["C", 67, 67, 72, 65, 75, 62],
-                ["L", 70, 72],
-                ["C", 59, 75, 46, 75, 36, 70],
-                ["Z"]
-            ],
-            fill: "skinMid",
-            stroke: "darkStroke",
-            strokeWidth: 2.5,
-            lineJoin: "round"
-        },
-
-        // Carapace top plate
-        {
-            type: "path",
-            commands: [
-                ["M", 34, 54],
-                ["C", 43, 45, 55, 41, 67, 43],
-                ["C", 73, 44, 78, 47, 81, 51],
-                ["L", 72, 58],
-                ["C", 62, 55, 50, 55, 39, 60],
-                ["Z"]
-            ],
-            fill: "carapaceFill",
-            stroke: "darkStroke",
-            strokeWidth: 3,
-            lineJoin: "round"
-        },
-
-        // Dorsal spikes
-        {
-            type: "path",
-            commands: [
-                ["M", 38, 49],
-                ["L", 35, 36],
-                ["L", 44, 46],
-                ["Z"],
-
-                ["M", 50, 44],
-                ["L", 49, 29],
-                ["L", 57, 42],
-                ["Z"],
-
-                ["M", 63, 45],
-                ["L", 67, 31],
-                ["L", 69, 46],
-                ["Z"]
-            ],
-            fill: "carapaceShade",
-            stroke: "darkStroke",
-            strokeWidth: 2.5,
-            lineJoin: "round"
-        },
-
-        // Head
-        {
-            type: "path",
-            commands: [
-                ["M", 61, 46],
-                ["C", 72, 40, 82, 41, 89, 47],
-                ["C", 92, 50, 92, 54, 88, 57],
-                ["C", 82, 62, 72, 62, 63, 58],
-                ["Z"]
-            ],
-            fill: "skinShade",
-            stroke: "darkStroke",
-            strokeWidth: 3.5,
-            lineJoin: "round"
-        },
-
-        // Upper jaw highlight / face plate
-        {
-            type: "path",
-            commands: [
-                ["M", 68, 48],
-                ["C", 75, 45, 82, 45, 87, 49],
-                ["L", 81, 53],
-                ["C", 77, 51, 72, 51, 68, 53],
-                ["Z"]
-            ],
-            fill: "carapaceFill",
-            stroke: "darkStroke",
-            strokeWidth: 2,
-            lineJoin: "round"
-        },
-
-        // Eye
-        {
-            type: "circle",
-            cx: 78,
-            cy: 50,
-            r: 2.8,
-            fill: "eyeFill",
-            stroke: "darkStroke",
-            strokeWidth: 1.5
-        },
-
-        // Front arm / main scythe
-        {
-            type: "path",
-            commands: [
-                ["M", 60, 61],
-                ["C", 66, 62, 72, 66, 74, 72],
-                ["L", 86, 78],
-                ["L", 78, 83],
-                ["L", 69, 75],
-                ["C", 66, 73, 63, 69, 60, 65],
-                ["Z"]
-            ],
-            fill: "skinMid",
-            stroke: "darkStroke",
-            strokeWidth: 3,
-            lineJoin: "round"
-        },
-
-        // Front scythe claw tip
-        {
-            type: "path",
-            commands: [
-                ["M", 84, 77],
-                ["L", 93, 79],
-                ["L", 82, 84],
-                ["Z"]
-            ],
-            fill: "clawFill",
-            stroke: "darkStroke",
-            strokeWidth: 2,
-            lineJoin: "round"
-        },
-
-        // Lower forelimb / secondary claw
-        {
-            type: "path",
-            commands: [
-                ["M", 53, 63],
-                ["C", 58, 66, 61, 71, 60, 76],
-                ["L", 52, 80],
-                ["L", 49, 73],
-                ["L", 49, 65],
-                ["Z"]
-            ],
-            fill: "skinShade",
-            stroke: "darkStroke",
-            strokeWidth: 2.5,
-            lineJoin: "round"
-        },
-
-        // Secondary claw tip
-        {
-            type: "path",
-            commands: [
-                ["M", 51, 79],
-                ["L", 47, 87],
-                ["L", 56, 82],
-                ["Z"]
-            ],
-            fill: "clawFill",
-            stroke: "darkStroke",
-            strokeWidth: 2,
-            lineJoin: "round"
-        },
-
-        // Little leg separation accents
-        {
-            type: "path",
-            commands: [
-                ["M", 35, 67],
-                ["L", 40, 79],
-                ["M", 44, 66],
-                ["L", 48, 76]
-            ],
-            stroke: "darkStroke",
-            strokeWidth: 2.5,
-            lineCap: "round"
-        }
-    ]
-};
-
-export const terranUnits = {
-    marine: terranMarine
-};
-
-export const zergUnits = {
-    zergling: zergZergling
-};
+import { zergUnits } from "./zergIcons";
 
 const definitions: UnitIconDefinition[] = [
     terranUnits.marine,
@@ -586,6 +15,8 @@ const definitions: UnitIconDefinition[] = [
 ];
 
 const aliases = new Map<string, UnitIconDefinition>();
+const svgCache = new Map<string, string>();
+const tokenCache = new Map<string, Record<string, string>>();
 
 for (const definition of definitions) {
     for (const alias of definition.aliases) {
@@ -600,11 +31,42 @@ export const unitIconCatalog = {
 
     render(ctx: CanvasContext, definition: UnitIconDefinition, options: UnitIconRenderOptions): void {
         renderIcon(ctx, definition, options);
+    },
+
+    toSvg(definition: UnitIconDefinition, options: UnitIconRenderOptions): string {
+        return toSvg(definition, options);
+    },
+
+    hydrateUnitIcons(root: ParentNode = document): void {
+        hydrateUnitIcons(root);
     }
 };
 
+export function hydrateUnitIcons(root: ParentNode = document): void {
+    const hosts = root.querySelectorAll<HTMLElement>("[data-unit-icon]");
+
+    for (const host of hosts) {
+        const commander = host.dataset.unitCommander ?? "";
+        const unitName = host.dataset.unitIcon ?? "";
+        const size = normalizeSize(Number(host.dataset.unitSize ?? 20));
+        const teamId = Number(host.dataset.teamId ?? 0);
+        const teamColor = host.dataset.teamColor || colorForTeam(teamId);
+        const definition = unitIconCatalog.resolve(commander, unitName);
+        const renderKey = `${commander}|${unitName}|${size}|${teamColor ?? ""}|${definition?.id ?? ""}`;
+
+        if (host.dataset.renderedIconKey === renderKey) {
+            continue;
+        }
+
+        host.dataset.renderedIconKey = renderKey;
+        host.innerHTML = definition
+            ? toSvg(definition, { size, teamColor })
+            : fallbackSvg(size, teamColor ?? "#8a949e");
+    }
+}
+
 function renderIcon(ctx: CanvasContext, definition: UnitIconDefinition, options: UnitIconRenderOptions): void {
-    const size = Math.max(1, options.size ?? 24);
+    const size = normalizeSize(options.size ?? 24);
     const x = options.x ?? size / 2;
     const y = options.y ?? size / 2;
     const scaleX = size / definition.viewBox.width;
@@ -676,19 +138,148 @@ function applyPath(ctx: CanvasContext, commands: UnitIconPathCommand[]): void {
     }
 }
 
+function toSvg(definition: UnitIconDefinition, options: UnitIconRenderOptions): string {
+    const size = normalizeSize(options.size ?? 20);
+    const teamColor = options.teamColor ?? "";
+    const cacheKey = `${definition.id}|${size}|${teamColor}`;
+    const cached = svgCache.get(cacheKey);
+    if (cached !== undefined) {
+        return cached;
+    }
+
+    const tokens = resolveTokens(definition, options.teamColor);
+    const layers = definition.layers.map(layer => toSvgLayer(layer, tokens)).join("");
+    const svg = `<svg xmlns="http://www.w3.org/2000/svg" width="${size}" height="${size}" viewBox="0 0 ${definition.viewBox.width} ${definition.viewBox.height}" role="img" aria-hidden="true" focusable="false">${layers}</svg>`;
+    svgCache.set(cacheKey, svg);
+    return svg;
+}
+
+function toSvgLayer(layer: UnitIconLayer, tokens: Record<string, string>): string {
+    const common = [
+        attr("fill", layer.fill ? resolvePaint(layer.fill, tokens) : undefined),
+        attr("stroke", layer.stroke ? resolvePaint(layer.stroke, tokens) : undefined),
+        attr("stroke-width", layer.strokeWidth),
+        attr("opacity", layer.opacity === undefined || layer.opacity === 1 ? undefined : layer.opacity),
+        layer.type === "path" ? attr("stroke-linecap", layer.lineCap) : "",
+        layer.type === "path" ? attr("stroke-linejoin", layer.lineJoin) : ""
+    ].join("");
+
+    if (layer.type === "circle") {
+        return `<circle cx="${layer.cx}" cy="${layer.cy}" r="${layer.r}"${common}/>`;
+    }
+
+    return `<path d="${commandsToPath(layer.commands)}"${common}/>`;
+}
+
+function commandsToPath(commands: UnitIconPathCommand[]): string {
+    return commands.map(command => command.join(" ")).join(" ");
+}
+
+function fallbackSvg(size: number, color: string): string {
+    const safeColor = escapeAttribute(color);
+    return `<svg xmlns="http://www.w3.org/2000/svg" width="${size}" height="${size}" viewBox="0 0 20 20" role="img" aria-hidden="true" focusable="false"><circle cx="10" cy="10" r="7" fill="${safeColor}99" stroke="${safeColor}" stroke-width="2"/></svg>`;
+}
+
 function resolveTokens(definition: UnitIconDefinition, teamColor: string | undefined): Record<string, string> {
     if (!teamColor) {
         return definition.tokens;
     }
 
-    return {
+    const cacheKey = `${definition.id}|${teamColor}`;
+    const cached = tokenCache.get(cacheKey);
+    if (cached !== undefined) {
+        return cached;
+    }
+
+    const palette = createTeamPalette(teamColor);
+    const tokens = {
         ...definition.tokens,
-        badgeFill: teamColor
+        badgeFill: teamColor,
+        badgeGlow: palette.light,
+        badgeStroke: palette.light,
+        armorFill: palette.light,
+        armorMid: teamColor,
+        armorShade: palette.dark,
+        armorDark: palette.deeper,
+        rifleFill: palette.deeper,
+        skinFill: teamColor,
+        skinMid: palette.dark,
+        skinShade: palette.deeper,
+        carapaceFill: palette.light,
+        carapaceShade: palette.dark
     };
+
+    tokenCache.set(cacheKey, tokens);
+    return tokens;
+}
+
+function createTeamPalette(color: string): { light: string; mid: string; dark: string; deeper: string } {
+    return {
+        light: mixHex(color, "#ffffff", 0.55),
+        mid: mixHex(color, "#ffffff", 0.18),
+        dark: mixHex(color, "#000000", 0.24),
+        deeper: mixHex(color, "#000000", 0.48)
+    };
+}
+
+function mixHex(left: string, right: string, weightRight: number): string {
+    const leftRgb = parseHexColor(left);
+    const rightRgb = parseHexColor(right);
+    if (!leftRgb || !rightRgb) {
+        return left;
+    }
+
+    const weightLeft = 1 - weightRight;
+    return toHex(
+        Math.round(leftRgb.r * weightLeft + rightRgb.r * weightRight),
+        Math.round(leftRgb.g * weightLeft + rightRgb.g * weightRight),
+        Math.round(leftRgb.b * weightLeft + rightRgb.b * weightRight));
+}
+
+function parseHexColor(color: string): { r: number; g: number; b: number } | null {
+    const match = /^#?([0-9a-f]{6})$/i.exec(color.trim());
+    if (!match) {
+        return null;
+    }
+
+    const value = Number.parseInt(match[1], 16);
+    return {
+        r: (value >> 16) & 255,
+        g: (value >> 8) & 255,
+        b: value & 255
+    };
+}
+
+function toHex(r: number, g: number, b: number): string {
+    return `#${hexByte(r)}${hexByte(g)}${hexByte(b)}`;
+}
+
+function hexByte(value: number): string {
+    return Math.max(0, Math.min(255, value)).toString(16).padStart(2, "0").toUpperCase();
 }
 
 function resolvePaint(paint: string, tokens: Record<string, string>): string {
     return tokens[paint] ?? paint;
+}
+
+function colorForTeam(teamId: number): string | undefined {
+    return TEAM_COLORS[teamId];
+}
+
+function attr(name: string, value: string | number | undefined): string {
+    return value === undefined ? "" : ` ${name}="${escapeAttribute(String(value))}"`;
+}
+
+function escapeAttribute(value: string): string {
+    return value
+        .replaceAll("&", "&amp;")
+        .replaceAll('"', "&quot;")
+        .replaceAll("<", "&lt;")
+        .replaceAll(">", "&gt;");
+}
+
+function normalizeSize(value: number): number {
+    return Number.isFinite(value) && value > 0 ? value : 20;
 }
 
 function getAliasKey(commander: string, unitName: string): string {
